+++ src/main/kotlin/kr/hhplus/be/server/interfaces/api/point/PointResponse.kt
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.interfaces.point
+package kr.hhplus.be.server.interfaces.api.point
 
 import kr.hhplus.be.server.application.point.PointResult
