+++ src/main/kotlin/kr/hhplus/be/server/interfaces/swagger/PointApi.kt
@@ -8,8 +8,8 @@ import io.swagger.v3.oas.annotations.parameters.RequestBody
 import io.swagger.v3.oas.annotations.responses.ApiResponse
 import io.swagger.v3.oas.annotations.responses.ApiResponses
 import io.swagger.v3.oas.annotations.tags.Tag
-import kr.hhplus.be.server.interfaces.point.PointRequest
-import kr.hhplus.be.server.interfaces.point.PointResponse
+import kr.hhplus.be.server.interfaces.api.point.PointRequest
+import kr.hhplus.be.server.interfaces.api.point.PointResponse
 
 @Tag(name = "Point API", description = "포인트 API")
 interface PointApi {
