+++ src/main/kotlin/kr/hhplus/be/server/interfaces/swagger/ProductApi.kt
@@ -7,7 +7,7 @@ import io.swagger.v3.oas.annotations.media.Schema
 import io.swagger.v3.oas.annotations.responses.ApiResponse
 import io.swagger.v3.oas.annotations.responses.ApiResponses
 import io.swagger.v3.oas.annotations.tags.Tag
-import kr.hhplus.be.server.interfaces.product.ProductResponse
+import kr.hhplus.be.server.interfaces.api.product.ProductResponse
 
 @Tag(name = "Product API", description = "상품 API")
 interface ProductApi {
