+++ src/test/java/kr/hhplus/be/server/interfaces/api/product/ProductControllerTest.kt
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.interfaces.product
+package kr.hhplus.be.server.interfaces.api.product
 
 import kr.hhplus.be.server.domain.product.ProductService
 import kr.hhplus.be.server.fixture.product.ProductInfoFixture
